Bug 9493: confirmations for user holds in opac are not translatable
In opac-user.tt, javascript for user holds confirmations (cancel, suspend, resume) are not translatable.
This is because text is in onclick attribute, withtout javascript method for translatable strings : _().

This patch moves those strings as variables into script tag and sets them translatable.

Test plan :
- set a hold for a user
- Go to OPAC user page "my summary"
- click on "Cancel"
=> you get a confirmation message
- click on "Suspend all holds"
=> you get a confirmation message
- click on "Resume all suspended holds"
=> you get a confirmation message
Test this is 'en' and another language.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>