Bug 7507 - Audience Subtype limit on OPAC Advanced Search does not have all Audiences
The OPAC audience subtype limit did not have all the audiences defined
by the MARC record (and available in the staff client advanced search).
It also incorrectly labelled some of the audiences that it did include.

This patch copies the (correct) audiences from the staff client template
to the OPAC template.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Test plan:
 * Search before applying by audience
 * Apply patch
 * Search by a newly appeared audience
 * Search by an audience that was there before

All tests pass!

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Replace tab with 4 spaces.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>