Bug 9711: Add css style for OPACPopupAuthorsSearch
Bug 5888 introduces a new syspref OPACPopupAuthorsSearch. If enabled, a
popup is displayed on the OPAC detail page.
This development does not add the css for the ccsr theme.

Test plan:
Enable the OPACPopupAuthorsSearch.
Check that the appearance is the same of prog and ccsr themes.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Tested on several authorities (authors, subjects, etc)
works as expected

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Test plan:
1) Switch to CCSR theme
2) Turn off XSLT views (leaving prefs empty)
3) Turn on OPACPopupAuthorsSearch
> Confirm layout is ok/broken in OPAC
4) Switch back to prog theme
> Confirm layout is ok/broken in OPAC
5) Switch off OPACPopupAuthorsSearch
> Confirm layout is ok/broken in OPAC
6) Switch back to CCSR
> Confirm layout is ok/broken in OPAC

Done.

Note: Feature only works with normal view (no XSLT)
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>