Bug 9813 - Update About page with license information about Bootstrap and jQuery
This patch adds license information about Bootstrap, jQuery, and
jQueryUI to the About page in the staff client.

To test, view the About page and confirm that information appears about
Bootstrap, jQuery, and jQueryUI with working links.

Fixed a small typo and changed erroneous h3's to h2's to match the other listings
Moved glyphicons link to indicate that it is licensed under the same terms as bootstrap - with Apache 2.0 license and not CC-BY.

Signed-off-by: Pierre Angot <tredok.pierre@gmail.com>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>